backend/requirements.txt
Django==5.2.12
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
